+++ strands-py/src/strands/experimental/bidi/io/__init__.py
@@ -1,7 +1,6 @@
 """IO channel implementations for bidirectional streaming."""
 
-from ..types.io import AudioProcessingConfig
 from .audio import BidiAudioIO
 from .text import BidiTextIO
 
-__all__ = ["AudioProcessingConfig", "BidiAudioIO", "BidiTextIO"]
+__all__ = ["BidiAudioIO", "BidiTextIO"]
