+++ strands-py/src/strands/experimental/bidi/io/audio.py
@@ -5,10 +5,10 @@ the output buffer is cleared to stop playback.
 
 Audio configuration is provided by the model via agent.model.config["audio"].
 
-Audio processing (acoustic echo cancellation, noise suppression, and automatic gain control) is available when
-pywebrtc-audio is installed (pip install strands-agents[bidi-aec]). Pass an ``AudioProcessingConfig`` to
-``BidiAudioIO`` to enable it. When enabled, the agent's speaker output is used as a reference signal to cancel echo
-from the microphone input so the model does not hear its own voice.
+Optional microphone audio processing (acoustic echo cancellation, noise suppression, and automatic gain
+control) is enabled by passing an ``AudioProcessorConfig`` (from ``strands.experimental.bidi.audio``) to
+``BidiAudioIO``. It requires pywebrtc-audio (pip install strands-agents[bidi-aec]); see that module for the
+processing implementation.
 """
 
 import asyncio
@@ -20,34 +20,13 @@ from typing import TYPE_CHECKING, Any
 import pyaudio
 
 from ..types.events import BidiAudioInputEvent, BidiAudioStreamEvent, BidiInterruptionEvent, BidiOutputEvent
-from ..types.io import AudioProcessingConfig, BidiInput, BidiOutput
+from ..types.io import BidiInput, BidiOutput
 
 if TYPE_CHECKING:
-    import numpy as np
-    import numpy.typing as npt
-
     from ..agent.agent import BidiAgent
+    from ..audio import AudioProcessorConfig, _AudioProcessor
 
 logger = logging.getLogger(__name__)
-
-_SUPPORTED_SAMPLE_RATES = (16000, 32000, 48000)
-"""Sample rates supported by pywebrtc-audio's AudioProcessor."""
-
-_FRAME_DURATION_MS = 10
-"""WebRTC audio processing operates on 10ms frames."""
-
-_DEFAULT_MAX_REF_FRAMES = 100
-"""Reference/mic buffer bound in frames (~1s at 10ms/frame). Both buffers share this bound so that under a
-sustained input stall they evict oldest data in lockstep, keeping the newest mic frame paired with the newest
-reference frame instead of letting the pairing drift apart."""
-
-_NOISE_SUPPRESSION = True
-"""Noise suppression is always applied when audio processing is enabled. Disabling it is an expert-only
-tuning case not exposed on the public config; promote to an AudioProcessingConfig field if real demand arises."""
-
-_AUTO_GAIN_CONTROL = True
-"""Automatic gain control is always applied when audio processing is enabled. Disabling it is an expert-only
-tuning case not exposed on the public config; promote to an AudioProcessingConfig field if real demand arises."""
 
 
 def _check_pywebrtc_available() -> None:
@@ -62,188 +41,6 @@ def _check_pywebrtc_available() -> None:
         raise ImportError(
             "pywebrtc-audio is required for audio processing. Install it with: pip install strands-agents[bidi-aec]"
         ) from error
-
-
-class _AudioProcessor:
-    """Owns microphone audio processing: reference buffer, WebRTC processor, and resampling.
-
-    Applies noise suppression and automatic gain control, and — when echo cancellation is enabled — cancels
-    the agent's own speaker audio from the mic input. A single instance is shared between the output and input
-    streams of a ``BidiAudioIO``: the output stream records every played frame as the far-end reference via
-    ``record_playback``, and the input stream processes captured mic audio via ``process_capture``.
-
-    The underlying pywebrtc-audio AudioProcessor is not thread-safe. It is only ever invoked from
-    ``process_capture`` on the input path, and those calls are serialized by the agent's ``await`` on the
-    executor, so no two ``process()`` calls overlap even though ``asyncio.to_thread`` uses a thread pool. The
-    reference buffer is a thread-safe queue shared across the input and output PyAudio callback threads.
-    """
-
-    def __init__(self, config: AudioProcessingConfig, max_ref_frames: int = _DEFAULT_MAX_REF_FRAMES) -> None:
-        """Initialize the processor.
-
-        Args:
-            config: Audio processing configuration.
-            max_ref_frames: Maximum number of reference frames to retain before dropping oldest.
-        """
-        self._config = config
-        self._max_ref_frames = max_ref_frames
-        self._ref_buffer: queue.Queue[bytes] = queue.Queue(maxsize=max_ref_frames)
-        self._processor: Any = None
-        self._input_rate: int | None = None
-        self._output_rate: int | None = None
-
-    def configure(self, input_rate: int, output_rate: int, num_channels: int) -> None:
-        """Build the underlying WebRTC processor for the given audio format.
-
-        Args:
-            input_rate: Microphone sample rate in Hz. Must be a supported rate.
-            output_rate: Speaker sample rate in Hz (reference is resampled to input_rate if different).
-            num_channels: Number of audio channels.
-
-        Raises:
-            ImportError: If pywebrtc-audio is not installed.
-            ValueError: If input_rate is not supported by pywebrtc-audio.
-        """
-        if input_rate not in _SUPPORTED_SAMPLE_RATES:
-            raise ValueError(
-                f"input_rate=<{input_rate}> | audio processing supports sample rates "
-                f"{_SUPPORTED_SAMPLE_RATES}. Configure the model's audio input_rate accordingly."
-            )
-
-        from pywebrtc_audio import AudioProcessor
-
-        # Discard any reference frames left over from a previous session so a restart never pairs stale
-        # far-end audio with fresh mic input. Safe without a lock: configure() runs before the PyAudio
-        # stream and processing threads exist.
-        _drain(self._ref_buffer)
-
-        self._input_rate = input_rate
-        self._output_rate = output_rate
-        self._processor = AudioProcessor(
-            sample_rate=input_rate,
-            num_channels=num_channels,
-            echo_cancellation=self._config.echo_cancellation,
-            noise_suppression=_NOISE_SUPPRESSION,
-            auto_gain_control=_AUTO_GAIN_CONTROL,
-            stream_delay_ms=self._config.stream_delay_ms,
-        )
-        logger.debug(
-            "input_rate=<%d>, output_rate=<%d>, channels=<%d> | audio processor started",
-            input_rate,
-            output_rate,
-            num_channels,
-        )
-
-    def record_playback(self, frame: bytes) -> None:
-        """Record a played speaker frame as the far-end reference.
-
-        Called from the output stream callback at the moment audio exits the speaker — the correct temporal
-        alignment point for echo cancellation. The frame is resampled to the input rate when the speaker and
-        mic rates differ. Drops the oldest reference frame on overflow.
-
-        No-op when echo cancellation is disabled, since no reference signal is needed.
-
-        Args:
-            frame: PCM int16 speaker audio at the output sample rate.
-        """
-        if not self._config.echo_cancellation:
-            return
-
-        frame = self._resample_reference(frame)
-        if self._ref_buffer.full():
-            logger.debug("ref_buffer_full=<True> | echo reference overflow, dropping oldest frame")
-            try:
-                self._ref_buffer.get_nowait()
-            except queue.Empty:
-                pass
-        self._ref_buffer.put_nowait(frame)
-
-    def process_capture(self, mic_data: bytes) -> bytes:
-        """Process captured mic audio: noise suppression, gain control, and echo cancellation if enabled.
-
-        Args:
-            mic_data: PCM int16 microphone audio.
-
-        Returns:
-            Cleaned PCM int16 audio of the same length. Empty input is returned unchanged (the WebRTC
-            processor rejects empty frames, and ``_BidiAudioBuffer.stop`` emits an empty shutdown sentinel).
-        """
-        import numpy as np_
-
-        if not mic_data:
-            return mic_data
-
-        near = np_.frombuffer(mic_data, dtype=np_.int16)
-        # With echo cancellation off there is no reference signal; pass far=None (noise suppression and
-        # auto gain control still run on the mic signal alone).
-        far = self._take_reference(near) if self._config.echo_cancellation else None
-        cleaned: npt.NDArray[np.int16] = self._processor.process(near, far)
-        return cleaned.astype(np_.int16).tobytes()
-
-    def clear_reference(self) -> None:
-        """Drain the reference buffer. Called on interruption to preserve time alignment.
-
-        The AEC filter itself is intentionally left converged — the acoustic path is unchanged by a barge-in.
-        """
-        _drain(self._ref_buffer)
-
-    def _take_reference(self, mic: "npt.NDArray[np.int16]") -> "npt.NDArray[np.int16]":
-        """Pull a reference frame aligned to the mic frame.
-
-        The WebRTC AudioProcessor requires a non-null far-end frame of the same length as the near-end
-        (mic) frame when echo cancellation is enabled. Any shortfall — including an empty reference buffer
-        because the speaker is silent — is filled with zeros (silence).
-
-        Args:
-            mic: int16 numpy array of mic samples for the current frame.
-
-        Returns:
-            int16 numpy array of reference samples, same length as ``mic``.
-        """
-        import numpy as np_
-
-        byte_count = mic.nbytes
-        data = bytearray()
-        while len(data) < byte_count:
-            try:
-                data.extend(self._ref_buffer.get_nowait())
-            except queue.Empty:
-                break
-
-        if len(data) < byte_count:
-            data.extend(b"\x00" * (byte_count - len(data)))
-
-        ref: npt.NDArray[np.int16] = np_.frombuffer(bytes(data[:byte_count]), dtype=np_.int16)
-        return ref
-
-    def _resample_reference(self, data: bytes) -> bytes:
-        """Resample speaker audio to the input rate via linear interpolation.
-
-        Returns data unchanged when rates match.
-        """
-        if self._output_rate is None or self._input_rate is None or self._output_rate == self._input_rate:
-            return data
-
-        import numpy as np_
-
-        samples = np_.frombuffer(data, dtype=np_.int16)
-        if len(samples) == 0:
-            return data
-
-        ratio = self._input_rate / self._output_rate
-        new_length = max(int(round(len(samples) * ratio)), 1)
-        positions = np_.linspace(0, len(samples) - 1, new_length)
-        resampled = np_.interp(positions, np_.arange(len(samples)), samples.astype(np_.float32))
-        return resampled.astype(np_.int16).tobytes()
-
-
-def _drain(buffer: "queue.Queue[Any]") -> None:
-    """Remove all items from a queue without blocking."""
-    while True:
-        try:
-            buffer.get_nowait()
-        except queue.Empty:
-            break
 
 
 class _BidiAudioBuffer:
@@ -353,7 +150,7 @@ class _BidiAudioInput(BidiInput):
     _DEVICE_INDEX = None
     _FRAMES_PER_BUFFER = 512
 
-    def __init__(self, config: dict[str, Any], processor: _AudioProcessor | None = None) -> None:
+    def __init__(self, config: dict[str, Any], processor: "_AudioProcessor | None" = None) -> None:
         """Extract configs.
 
         Args:
@@ -405,7 +202,7 @@ class _BidiAudioInput(BidiInput):
             # 10ms preserves AEC quality (ERLE is unchanged at 20/30/40ms), so the wire cadence could be
             # decoupled by batching several processed frames into one event (e.g. 20-30ms). Deferred until
             # the cadence is measured against a live provider for throttling/cost impact.
-            frames_per_buffer = self._rate * _FRAME_DURATION_MS // 1000
+            frames_per_buffer = self._rate * self._processor._FRAME_DURATION_MS // 1000
         else:
             frames_per_buffer = self._configured_frames_per_buffer
 
@@ -472,7 +269,7 @@ class _BidiAudioOutput(BidiOutput):
     _DEVICE_INDEX = None
     _FRAMES_PER_BUFFER = 512
 
-    def __init__(self, config: dict[str, Any], processor: _AudioProcessor | None = None) -> None:
+    def __init__(self, config: dict[str, Any], processor: "_AudioProcessor | None" = None) -> None:
         """Extract configs.
 
         Args:
@@ -502,7 +299,7 @@ class _BidiAudioOutput(BidiOutput):
             # is measured in samples at the stream's own (output) rate, so sizing off input_rate would open a
             # buffer of the wrong duration when output_rate != input_rate (e.g. Gemini 16k in / 24k out),
             # yielding short reference frames that get zero-padded and degrade echo cancellation.
-            frames_per_buffer = self._rate * _FRAME_DURATION_MS // 1000
+            frames_per_buffer = self._rate * self._processor._FRAME_DURATION_MS // 1000
         else:
             frames_per_buffer = self._configured_frames_per_buffer
 
@@ -567,46 +364,49 @@ class BidiAudioIO:
     Reads microphone audio via ``input()`` and plays agent audio via ``output()``. On interruption, the
     playback buffer is cleared to stop the agent mid-response.
 
-    When an ``AudioProcessingConfig`` is provided, the microphone signal gets noise suppression and automatic
-    gain control, and (when echo cancellation is enabled) the agent's speaker output is used as a reference to
-    cancel echo from the mic input, preventing the model from hearing its own voice. The same processor
-    instance is shared between the input and output channels this factory produces, so echo cancellation only
-    works when both ``input()`` and ``output()`` come from the *same* ``BidiAudioIO`` instance.
+    When an ``AudioProcessorConfig`` is passed as ``processor``, the microphone signal gets noise suppression
+    and automatic gain control, and (when echo cancellation is enabled) the agent's speaker output is used as a
+    reference to cancel echo from the mic input, preventing the model from hearing its own voice. The same
+    processor instance is shared between the input and output channels this factory produces, so echo
+    cancellation only works when both ``input()`` and ``output()`` come from the *same* ``BidiAudioIO``
+    instance.
 
     Audio processing requires pywebrtc-audio (``pip install strands-agents[bidi-aec]``) and a microphone
     sample rate of 16000, 32000, or 48000 Hz (set via the model's audio config).
 
     Example:
         ```python
-        from strands.experimental.bidi.io import AudioProcessingConfig, BidiAudioIO
+        from strands.experimental.bidi.audio import AudioProcessorConfig
+        from strands.experimental.bidi.io import BidiAudioIO
 
         # Plain mic/speaker, no processing (a headset is recommended to avoid echo):
         audio_io = BidiAudioIO()
         await agent.run(inputs=[audio_io.input()], outputs=[audio_io.output()])
 
         # Full processing: echo cancellation, noise suppression, and auto gain control:
-        audio_io = BidiAudioIO(audio_processing=AudioProcessingConfig())
+        audio_io = BidiAudioIO(processor=AudioProcessorConfig())
         await agent.run(inputs=[audio_io.input()], outputs=[audio_io.output()])
 
         # Noise suppression and auto gain control without echo cancellation (e.g. headset users):
-        audio_io = BidiAudioIO(audio_processing=AudioProcessingConfig(echo_cancellation=False))
+        audio_io = BidiAudioIO(processor=AudioProcessorConfig(echo_cancellation=False))
         await agent.run(inputs=[audio_io.input()], outputs=[audio_io.output()])
 
         # Processing on a specific input device:
-        audio_io = BidiAudioIO(audio_processing=AudioProcessingConfig(), input_device_index=1)
+        audio_io = BidiAudioIO(processor=AudioProcessorConfig(), input_device_index=1)
         await agent.run(inputs=[audio_io.input()], outputs=[audio_io.output()])
         ```
     """
 
-    def __init__(self, *, audio_processing: AudioProcessingConfig | None = None, **config: Any) -> None:
+    def __init__(self, **config: Any) -> None:
         """Initialize audio devices.
 
         Args:
-            audio_processing: Enable microphone audio processing (noise suppression, automatic gain control,
-                and optionally echo cancellation) by passing an ``AudioProcessingConfig``. Requires
-                pywebrtc-audio (pip install strands-agents[bidi-aec]). Defaults to None (processing disabled).
-            **config: Optional device configuration:
+            **config: Optional configuration:
 
+                - processor (AudioProcessorConfig): Enable microphone audio processing (noise suppression,
+                  automatic gain control, and optionally echo cancellation) by passing an
+                  ``AudioProcessorConfig`` (from ``strands.experimental.bidi.audio``). Requires pywebrtc-audio
+                  (pip install strands-agents[bidi-aec]). Defaults to None (processing disabled).
                 - input_buffer_size (int): Maximum input buffer size (default: None). Ignored when echo
                   cancellation is on — the mic buffer is bound to the reference horizon so the two stay aligned.
                 - input_device_index (int): Specific input device (default: None = system default)
@@ -616,13 +416,16 @@ class BidiAudioIO:
                 - output_frames_per_buffer (int): Output buffer size (default: 512, ignored when processing is on)
 
         Raises:
-            ImportError: If audio_processing is set but pywebrtc-audio is not installed.
+            ImportError: If a processor config is set but pywebrtc-audio is not installed.
         """
+        processor_config: AudioProcessorConfig | None = config.pop("processor", None)
         self._config = config
 
-        if audio_processing is not None:
+        if processor_config is not None:
             _check_pywebrtc_available()
-            self._processor: _AudioProcessor | None = _AudioProcessor(audio_processing)
+            from ..audio import _AudioProcessor
+
+            self._processor: _AudioProcessor | None = _AudioProcessor(processor_config)
         else:
             self._processor = None
